Wandb

+++ src/utilities/metrics.py
@@ -53,9 +53,9 @@ def get_wasserstein_distance(predicted_label, true_label):
     return torch.from_numpy(np.array(dist/predicted_label.shape[0]))
 
 def get_classification_metrics(label_batch, prediction_batch, cutoff=0.05):
-    batch_size = label_batch.shape[0]
-    label_mask = (label_batch > cutoff).type(torch.int).float()
-    prediction_mask = (prediction_batch > cutoff).type(torch.int).float()
+    batch_size = float(label_batch.shape[0])
+    label_mask = (label_batch > cutoff).type(torch.int)
+    prediction_mask = (prediction_batch > cutoff).type(torch.int)
     fp = torch.sum(label_mask - prediction_mask < -0.1)/batch_size
     fn = torch.sum(label_mask - prediction_mask > 0.1)/batch_size
     tp = torch.sum(torch.einsum("bi,bi->b", prediction_mask, label_mask))
